backend/requirements.txt
asgiref==3.9.2
attrs==25.3.0
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.95.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.10.1
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
psycopg==3.1.19
psycopg-binary==3.1.19
PyJWT==2.10.1
python-dotenv==1.0.1
PyYAML==6.0.3
referencing==0.36.2
rpds-py==0.27.1
setuptools==80.9.0
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
wheel==0.45.1
